+++ util/string_tools.hpp
@@ -1,0 +1,21 @@
+#ifndef __STRING_TOOLS_HPP
+#define __STRING_TOOLS_HPP
+
+#include <algorithm>
+#include <string>
+#include <unordered_set>
+
+#include <zlib.h>
+
+#define IO_SUCCESS 0
+#define IO_FAILURE 1
+#define IO_ABORTED 2
+
+void strToUpper(std::string&);
+void splitStringUnique(std::unordered_set<std::string>&, std::string&, char, bool);
+void stringReplaceAll(std::string&, const std::string&, const std::string&);
+void trimString(std::string&);
+
+int deflateText(std::string&);
+
+#endif